Get proper classes from Arbor rather than TeachingGroups

+++ dev/arbor_import.php
@@ -59,6 +59,9 @@ if (empty($_GET['baseline_done'])) {
             // Does this Baseline already exist?  Overwrite if so.
             $details[Baseline::ID] = $baseline['id'];
             $details[Baseline::GRADE] = $baseline['grade']['displayName'];
+            if (is_null($details[Baseline::GRADE])) {
+                continue;
+            }
             $details[Baseline::STUDENT_ID] = $baseline['student']['id'];
             $details[Baseline::NAME] = $baseline['assessment']['displayName'];
             $details[Baseline::MIS_ASSESSMENT_ID] = $baseline['assessment']['id'];
@@ -71,58 +74,78 @@ if (empty($_GET['baseline_done'])) {
     die('<a href="?baseline_done=yes" class="btn btn-primary">Baselines done!  Now click to import groups</a>');
 }
 
-$ay = Config::academic_year;
+if (empty($_GET['year_page'])) {
+    $year_page = 0;
+} else {
+    $year_page = $_GET['year_page'];
+}
 
 $query = "query {
-  TeachingGroup (academicYear__code: \"$ay\") {
+  AcademicUnit (academicYear__code: \"" . Config::academic_year . "\" page_size: 100 page_num: $year_page) {
+    id
     displayName
-    memberships {
+    allMemberships {
       student {
+        id
         lastName
         firstName
-        id
       }
     }
   }
 }";
 
 $data = $client->rawQuery($query)->getData();
-  
-// Clear old memberships out
-(new Database())->dosql("DELETE FROM studentgroupmembership;");
 
-foreach ($data['TeachingGroup'] as $group) {
+if (empty($data['AcademicUnit'])) {
+    die("<div class=\"row\">Complete!</div>");
+}
+
+// Clear old memberships out
+if ($year_page == 0) {
+    (new Database())->dosql("DELETE FROM studentgroupmembership;");
+}
+
+$year_page++;
+echo "<div class=\"row\"><a href=\"?baseline_done=yes&year_page=$year_page\" class=\"btn btn-primary\">Now click for Page $year_page</a></div>";
+
+foreach ($data['AcademicUnit'] as $group) {
+    $displayNames = explode(":", $group['displayName']);
+    $displayName = end($displayNames);
     if (!empty($dGroup = TeachingGroup::retrieveByDetail(TeachingGroup::ID, $group['id']))) {
         $dGroup = $dGroup[0];
-        $dGroup->setName($group['displayName']);
+        $dGroup->setName($displayName);
         echo "<div class=\"row\">Scanned TeachingGroup: " . $dGroup->getName() . "</div>"; 
     } else {
-        $group[TeachingGroup::NAME] = $group['displayName'];
+        $group[TeachingGroup::NAME] = $displayName;
         $group[TeachingGroup::ACADEMIC_YEAR] = Config::academic_year;
         $dGroup = new TeachingGroup($group);
         echo "<div class=\"row\">New TeachingGroup: " . $dGroup->getName() . "</div>";
     }
     $dGroup->commit();
-    
-    foreach ($group['memberships'] as $membership) {
+    $numGroupMembers = 0;
+    foreach ($group['allMemberships'] as $membership) {
         if (!empty($dStudent = Student::retrieveByDetail(Student::ID, $membership['student']['id']))) {
             $dStudent = $dStudent[0];
             $dStudent->setNames($membership['student']['firstName'], $membership['student']['lastName']);
-            echo "<div class=\"row\">&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Scanned Student: " . $dStudent->getName();
+            // echo "<div class=\"row\">&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Scanned Student: " . $dStudent->getName();
         } else {
             $dStudent = new Student([
+                Student::ID         => $membership['student']['id'],
                 Student::FIRST_NAME => $membership['student']['firstName'],
                 Student::LAST_NAME  => $membership['student']['lastName'],
-                Student::ID         => $membership['student']['id']
             ]);
-            echo "<div class=\"row\">&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;New Student: " . $dStudent->getName();
+            echo "<div class=\"row\">&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;New Student: " . $dStudent->getName() . "</div>";
         }
         $dGroup->addMember($dStudent);
         
         $dStudent->commit();
         
-        echo "... added to db and membership made</div>";
+        $numGroupMembers++;
+        
+        // echo "... added to db and membership made</div>";
     }
+    
+    echo "<div class=\"row\">Group {$dGroup->getName()} now has $numGroupMembers members.</div>";
 }
 ?>
 	</div>
